+++ src/components/Card/Card.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import like from './../assets/like.svg';
-import bookmark from './../assets/bookmark.svg';
+import like from './../../assets/like.svg';
+import bookmark from './../../assets/bookmark.svg';
 
 class Card extends Component<CardProps> {
   state = {
@@ -24,7 +24,7 @@ class Card extends Component<CardProps> {
 
   render() {
     return (
-      <li className="card" aria-label="card">
+      <li className="card" data-testid="card">
         <div className="card__image-container">
           <img src={this.props.image} alt="card-image" className="card__image" />
         </div>
@@ -37,6 +37,7 @@ class Card extends Component<CardProps> {
               className="card__btn"
               onClick={this.hadnleBtnClick.bind(this)}
               data-btntype="likes"
+              data-testid="likes"
             >
               <img src={like} alt="like-icon" />
               <span> {this.state.likes}</span>
@@ -45,6 +46,7 @@ class Card extends Component<CardProps> {
               className="card__btn"
               onClick={this.hadnleBtnClick.bind(this)}
               data-btntype="bookmarks"
+              data-testid="bookmarks"
             >
               <img src={bookmark} alt="bookmark-icon" />
               <span>{this.state.bookmarks}</span>
