+++ src/components/Form/FormProps.tsx
@@ -1,0 +1,3 @@
+interface FormProps {
+  addCard: (card: CardProps) => void;
+}
