+++ src/pages/Formpage/Formpage.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import Form from './../components/Form';
-import Cards from './../components/Cards';
+import Form from '../../components/Form/Form';
+import Cards from '../../components/Cards/Cards';
 
 class FormPage extends Component {
   state = {
