+++ test/module02.test.tsx
@@ -1,26 +1,37 @@
 import React from 'react';
 
-import { describe, test, expect } from 'vitest';
-import { render, screen, fireEvent, queryByAttribute } from '@testing-library/react';
+import { describe, test, expect, beforeEach } from 'vitest';
+import { render, screen, fireEvent } from '@testing-library/react';
 
-import Cards from '../src/components/Cards';
-import Form from '../src/components/Form';
-import FormPage from './../src/pages/Formpage';
+import Cards from '../src/components/Cards/Cards';
+import Form from '../src/components/Form/Form';
+import FormPage from '../src/pages/Formpage/Formpage';
 
 describe('Search tests', () => {
+  beforeEach(() => {
+    render(<FormPage />);
+  });
+
   test('Should render Form component', () => {
-    render(<Form addCard={() => {}} />);
-    const form = screen.getByLabelText('form') as HTMLFormElement;
+    const form = screen.getByTestId('form') as HTMLFormElement;
     expect(form).toBeDefined();
   });
 
   test('Should render Form component', () => {
-    render(<Form addCard={() => {}} />);
-    const form = screen.getByLabelText('form') as HTMLFormElement;
-    expect(form).toBeDefined();
-  });
-});
+    const titleInput = screen.getByTestId('title');
+    expect(titleInput).toBeDefined();
+    fireEvent.input(titleInput, { target: { value: 'title' } });
+    expect(titleInput).toHaveValue('title');
 
-describe('Cards tests', () => {
-  console.log(1);
+    const priceInput = screen.getByTestId('price');
+    fireEvent.input(titleInput, { target: { value: 1000 } });
+
+    const descrInput = screen.getByTestId('price');
+    fireEvent.input(descrInput, { target: { value: 'descr' } });
+
+    const submitInput = screen.getByTestId('submit');
+    fireEvent.submit(submitInput);
+
+    expect(submitInput).toBeDisabled();
+  });
 });
